feat: should optionally clean the history
#9

+++ src/tab-cleaner.ts
@@ -1,6 +1,10 @@
+import {Options} from './storage';
+
 (async function () {
 
-    if ('localStorage' in window) {
+    const options : Options = window['CPDN_OPTIONS'];
+
+    if (options.scopes.localStorage && 'localStorage' in window) {
         try {
             console.info('clear localStorage');
             localStorage.clear();
@@ -9,7 +13,7 @@
         }
     }
 
-    if ('sessionStorage' in window) {
+    if (options.scopes.sessionStorage && 'sessionStorage' in window) {
         try {
             console.info('clear sessionStorage');
             sessionStorage.clear();
@@ -18,17 +22,17 @@
         }
     }
 
+    let indexedDB: any;
     if (!('indexedDB' in window)) {
-        // @ts-ignore
-        window.indexedDB = window.mozIndexedDB || window.webkitIndexedDB || window.msIndexedDB;
+        indexedDB = window['indexedDB'] || window[`mozIndexedDB`] || window[`webkitIndexedDB`] || window[`msIndexedDB`];
     }
 
-    if (!window.indexedDB) {
+    if (options.scopes.indexedDB && !indexedDB) {
         console.info('clear indexedDB', Object.keys(window.indexedDB));
         // It is not yet possible to discover related IndexDB database.
     }
 
-    if ('serviceWorker' in navigator) {
+    if (options.scopes.serviceWorker && 'serviceWorker' in navigator) {
         console.info('clear serviceWorker');
         try {
             const registrations = await navigator.serviceWorker.getRegistrations();
@@ -46,7 +50,7 @@
         }
     }
 
-    if ('caches' in window) {
+    if (options.scopes.caches && 'caches' in window) {
         console.info('clear caches');
         try {
             const keys = await caches.keys();
